renamed controller helper to user helper
